Move all tests related to native SDK transactions into their own package
It makes it easier to test just that functionality
while iterating.
